chore(combo): extract pure error predicates and quota status helpers to comboPredicates
